+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### [0.7.0](https://github.com/ElvenTools/elven-tools-sft-smart-contract/releases/tag/v0.7.0) (2024-02-10)
+- add giveaway functionality
+- update `multiversx-sc` libraries
+
 ### [0.6.0](https://github.com/ElvenTools/elven-tools-sft-smart-contract/releases/tag/v0.6.0) (2024-01-05)
 - add possibility to define special roles
 - update `multiversx-sc` libraries